requirements.txt
opencv-python
numpy
Hobot.GPIO
